change to new route format
