+++ packages/server/lib/controllers/oauth.controller.ts
@@ -686,185 +686,6 @@ class OAuthController {
         }
     }
 
-    public async oauth2OutboundRequest(req: Request, res: Response<any, Required<RequestLocals>>, next: NextFunction) {
-        const { environment, account, connectSession } = res.locals;
-        const { providerConfigKey } = req.params;
-        const receivedConnectionId = req.query['connection_id'] as string | undefined;
-        let connectionId = receivedConnectionId || connectionService.generateConnectionId();
-        const connectionConfig: ConnectionConfig = req.query['params'] != null ? getConnectionConfig(req.query['params']) : {};
-        const query = req.query;
-        const isConnectSession = res.locals['authType'] === 'connectSession';
-
-        if (query['installation'] !== 'outbound') {
-            errorManager.errRes(res, 'missing_client_id');
-
-            return;
-        }
-
-        if (isConnectSession && receivedConnectionId) {
-            errorRestrictConnectionId(res);
-            return;
-        }
-
-        let logCtx: LogContext | undefined;
-
-        try {
-            logCtx =
-                isConnectSession && connectSession.operationId
-                    ? logContextGetter.get({ id: connectSession.operationId, accountId: account.id })
-                    : await logContextGetter.create(
-                          {
-                              operation: { type: 'auth', action: 'create_connection' },
-                              meta: { authType: 'oauth2-oubound', connectSession: endUserToMeta(res.locals.endUser) },
-                              expiresAt: defaultOperationExpiration.auth()
-                          },
-                          { account, environment }
-                      );
-
-            if (!providerConfigKey) {
-                errorManager.errRes(res, 'missing_connection');
-
-                return;
-            }
-
-            if (!isConnectSession) {
-                const hmac = req.query['hmac'] as string | undefined;
-
-                const checked = await hmacCheck({ environment, logCtx, providerConfigKey, connectionId, hmac, res });
-                if (!checked) {
-                    return;
-                }
-            }
-
-            const config = await configService.getProviderConfig(providerConfigKey, environment.id);
-            if (!config) {
-                void logCtx.error('Unknown provider config');
-                await logCtx.failed();
-
-                errorManager.errRes(res, 'unknown_provider_config');
-
-                return;
-            }
-
-            const provider = getProvider(config.provider);
-            if (!provider) {
-                void logCtx.error('Unknown provider');
-                await logCtx.failed();
-                res.status(404).send({ error: { code: 'unknown_provider_template' } });
-                return;
-            }
-
-            if (provider.auth_mode !== 'OAUTH2') {
-                void logCtx.error('Provider does not support OAuth2 Outbound Installation creation', { provider: config.provider });
-                await logCtx.failed();
-
-                errorManager.errRes(res, 'invalid_auth_mode');
-
-                return;
-            }
-
-            if (!(await isIntegrationAllowed({ config, res, logCtx }))) {
-                return;
-            }
-
-            if (isConnectSession) {
-                // Reconnect mechanism
-                if (connectSession.connectionId) {
-                    const connection = await connectionService.getConnectionById(connectSession.connectionId);
-                    if (!connection) {
-                        void logCtx.error('Invalid connection');
-                        await logCtx.failed();
-                        res.status(400).send({ error: { code: 'invalid_connection' } });
-                        return;
-                    }
-                    connectionId = connection?.connection_id;
-                }
-            }
-
-            await logCtx.enrichOperation({ integrationId: config.id!, integrationName: config.unique_key, providerName: config.provider });
-
-            const updatedConnectionConfig: ConnectionConfig = {
-                ...connectionConfig,
-                pending: true,
-                pendingLog: logCtx.id
-            };
-
-            const [updatedConnection] = await connectionService.upsertConnection({
-                connectionId,
-                providerConfigKey,
-                parsedRawCredentials: { type: 'OAUTH2' } as any,
-                connectionConfig: updatedConnectionConfig,
-                environmentId: environment.id
-            });
-            if (!updatedConnection) {
-                res.status(500).send({ error: { code: 'server_error', message: 'failed to create connection' } });
-                void logCtx.error('Failed to create connection');
-                await logCtx.failed();
-                return;
-            }
-
-            if (isConnectSession) {
-                await linkConnection(db.knex, { endUserId: connectSession.endUserId, connection: updatedConnection.connection });
-            }
-
-            await logCtx.enrichOperation({ connectionId: updatedConnection.connection.id, connectionName: updatedConnection.connection.connection_id });
-            void logCtx.info('OAuth2 Outbound Installation creation was successful');
-            await logCtx.success();
-            void connectionCreatedHook(
-                {
-                    connection: updatedConnection.connection,
-                    environment,
-                    account,
-                    auth_mode: 'OAUTH2',
-                    operation: updatedConnection.operation,
-                    endUser: isConnectSession ? res.locals['endUser'] : undefined
-                },
-                account,
-                config,
-                logContextGetter
-            );
-
-            metrics.increment(metrics.Types.AUTH_SUCCESS, 1, { auth_mode: provider.auth_mode });
-
-            res.status(200).send({ providerConfigKey: providerConfigKey, connectionId: connectionId });
-        } catch (err) {
-            const prettyError = stringifyError(err, { pretty: true });
-
-            void connectionCreationFailedHook(
-                {
-                    connection: { connection_id: receivedConnectionId!, provider_config_key: providerConfigKey! },
-                    environment,
-                    account,
-                    auth_mode: 'OAUTH2',
-                    error: {
-                        type: 'unknown',
-                        description: `Error during Unauth create: ${prettyError}`
-                    },
-                    operation: 'unknown'
-                },
-                account
-            );
-            if (logCtx) {
-                void logCtx.error('Error during OAuth2 Outbound Installation credentials creation', { error: err });
-                await logCtx.failed();
-            }
-
-            errorManager.report(err, {
-                source: ErrorSourceEnum.PLATFORM,
-                operation: LogActionEnum.AUTH,
-                environmentId: environment.id,
-                metadata: {
-                    providerConfigKey,
-                    connectionId: receivedConnectionId
-                }
-            });
-
-            metrics.increment(metrics.Types.AUTH_FAILURE, 1, { auth_mode: 'OAUTH2' });
-
-            next(err);
-        }
-    }
-
     private async appRequest(
         provider: Provider,
         providerConfig: ProviderConfig,
